Remove from_coordinates_unchecked() API, introduce copy_point()
Now that the Sinsemilla and ECC instructions are in different
crates, ECC can no longer limit from_coordinates_unchecked()
to just the crate.

Instead of exposing this API publically, we introduce a copy_point()
instruction that allows the output of Sinsemilla hash_to_point to be
copied and checked by the ECC chip.